@@ -4,7 +4,7 @@ package com.example.filmography.dto;
 import lombok.Data;
 
 @Data
-public class AddFilmsDto {
+public class AddFilmDto {
         Long filmId;
         Long directorId;
     }
