@@ -5,6 +5,8 @@ import lombok.*;
 import javax.persistence.*;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.Date;
+import java.util.Set;
 
 @Entity
 @Table(name = "users")
@@ -12,16 +14,13 @@ import java.time.LocalDateTime;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-@SequenceGenerator(name = "default_generator", sequenceName = "users_seq", allocationSize = 1)
+@SequenceGenerator(name = "default_generator", sequenceName = "user_seq", allocationSize = 1)
 public class User extends GenericModel {
-    //  @Id
-    //  @Column(name = "id")
-    //   private Long id;
 
-    @Column(name = "login", nullable = false)
+    @Column(name = "login", unique = true)
     private String login;
 
-    @Column(name = "password", nullable = false)
+    @Column(name = "password")
     private String password;
 
     @Column(name = "first_name")
@@ -33,7 +32,7 @@ public class User extends GenericModel {
     @Column(name = "middle_name")
     private String middleName;
 
-    @Column(name = "birth_date")
+    @Column(name = "birthdate")
     private LocalDate birthDate;
 
     @Column(name = "phone")
@@ -42,37 +41,45 @@ public class User extends GenericModel {
     @Column(name = "address")
     private String address;
 
-    @Column(name = "email")
+    @Column(name = "email", unique = true)
     private String email;
 
+    @Column(name = "created_when")
+    private LocalDateTime createdWhen;
 
-    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL) // добавид каскад тип
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(
             name = "role_id",
-            foreignKey = @ForeignKey(name = "FK_USER_ROLES")
-    )
+            foreignKey = @ForeignKey(name = "FK_USER_ROLES"))
+    //@JsonIgnore
     private Role role;
+
+    @OneToMany(mappedBy = "user")
+    private Set<Order> orders;
+
+    @Column(name = "change_password_token")
+    private String changePasswordToken;
 
     @Override
     public String toString() {
         return "User{" +
-                "login='" + login + '\'' +
-                ", password='" + password + '\'' +
-                ", firstName='" + firstName + '\'' +
-                ", lastName='" + lastName + '\'' +
-                ", middleName='" + middleName + '\'' +
-                ", birthDate=" + birthDate +
-                ", phone='" + phone + '\'' +
-                ", address='" + address + '\'' +
-                ", email='" + email + '\'' +
-                ", role=" + role.getId() +
-                '}';
+               "login='" + login + '\'' +
+               ", password='" + password + '\'' +
+               ", firstName='" + firstName + '\'' +
+               ", lastName='" + lastName + '\'' +
+               ", middleName='" + middleName + '\'' +
+               ", birthDate=" + birthDate +
+               ", phone='" + phone + '\'' +
+               ", address='" + address + '\'' +
+               ", email='" + email + '\'' +
+               ", createdWhen=" + createdWhen +
+               ", role=" + role.getId() +
+               '}';
     }
 
     @Builder
-    public User(Long id, String login, String password, String firstName, String lastName,
-                String middleName, LocalDate birthDate, String phone, String address, String email, Role role) {
-        super(id);
+    public User(Long id, LocalDateTime createdWhen, String createdBy, LocalDateTime updatedWhen, String updatedBy, LocalDateTime deletedWhen, String deletedBy, boolean isDeleted, String login, String password, String firstName, String lastName, String middleName, LocalDate birthDate, String phone, String address, String email, Role role) {
+        super(id, createdWhen, createdBy, updatedWhen, updatedBy, deletedWhen, deletedBy, isDeleted);
         this.login = login;
         this.password = password;
         this.firstName = firstName;
@@ -83,8 +90,6 @@ public class User extends GenericModel {
         this.address = address;
         this.email = email;
         this.role = role;
-
     }
-
-   // public String toString(){}
 }
+
