@@ -1,6 +1,8 @@
 package com.example.filmography.repository;
 
 import com.example.filmography.model.Director;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,9 +10,16 @@ import java.util.List;
 import java.util.Set;
 
 @Repository
-public interface DirectorRepository extends JpaRepository <Director, Long>{
+public interface DirectorRepository extends GenericRepository<Director> {
+
     Set<Director> findAllByIdIn(Set<Long> ids);
 
-    List<Director> findAllByDirectorFIO(String directorFIO);
+    Page<Director> findAllByDirectorFIO(Pageable pageable, String directorFIO);
 
+    Page<Director> findAllByDirectorFIOContaining(Pageable pageable, String directorFIO);
+
+    Page<Director> findAllByDirectorFIOContainingAndIsDeletedFalse(Pageable pageable, String directorFIO);
+
+    Page<Director> findAllByIsDeletedFalse(Pageable pageable);
 }
+
