@@ -1,7 +1,11 @@
 package com.example.filmography.repository;
 
+
 import com.example.filmography.model.GenericModel;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.NoRepositoryBean;
 
-public interface GenericRepository < T extends GenericModel> extends JpaRepository< T, Long> {
+
+@NoRepositoryBean
+public interface GenericRepository<T extends GenericModel> extends JpaRepository<T, Long> {
 }
