@@ -1,63 +1,100 @@
 package com.example.filmography.service;
 
-import com.example.filmography.dto.AddFilmsDto;
-import com.example.filmography.dto.DirectorDto;
+import com.example.filmography.dto.AddFilmDto;
+
 import com.example.filmography.model.Director;
 import com.example.filmography.model.Film;
+import com.example.filmography.model.Order;
 import com.example.filmography.repository.DirectorRepository;
+import com.example.filmography.repository.FilmRepository;
+import com.example.filmography.repository.OrderRepository;
+import com.example.filmography.service.userDetails.CustomUserDetails;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import java.time.LocalDateTime;
+
+import java.util.Set;
 
 @Service
-public class DirectorService {
-    private final DirectorRepository directorRepository;
-    private final FilmService filmService;
+public class DirectorService extends GenericService<Director> {
 
+    private final DirectorRepository repository;
+    private final OrderRepository orderRepository;
+    private final FilmRepository filmRepository;
 
-    public DirectorService(DirectorRepository directorRepository, FilmService filmService) {
-        this.directorRepository = directorRepository;
-        this.filmService = filmService;
-
+    protected DirectorService(DirectorRepository repository, OrderRepository orderRepository,
+                              FilmRepository filmRepository) {
+        super(repository);
+        this.repository = repository;
+        this.orderRepository = orderRepository;
+        this.filmRepository = filmRepository;
     }
 
-    public List<Director> getList() {
-        return directorRepository.findAll();
-    }
-
-    public Director getOne(Long id) {
-        return directorRepository.findById(id).orElseThrow();
-    }
-
-    public Director create(Director director) {
-        return directorRepository.save(director);
-    }
-
-    public Director update(Director director) {
-        return directorRepository.save(director);
-    }
-
+    //Удалить участника фильма, если нет заказов с его фильмами,
+    // удалить фильмы, если он был единственным участником.
     public void delete(Long id) {
-        directorRepository.delete(directorRepository.findById(id).orElseThrow());
-    }
+        Set<Order> orders = orderRepository.findAllByDirectorId(id);
 
-
-    public Director addFilm(AddFilmsDto addFilmsDto) {
-        Director director = getOne(addFilmsDto.getDirectorId());
-        Film film = filmService.getOne(addFilmsDto.getFilmId());
-        director.getFilms().add(film);
-        //    film.getDirectors().add(director);
-        // film.getDirectors().add(director);
-        //   filmService.update(film);
-        return update(director);
-    }
-
-  /*  public DirectorDto getListWithFilmIds(AddFilmsDto addFilmsDto, DirectorDto directorDto) {
-        Director director = getOne(addFilmsDto.getDirectorId());
-        for( filmId : addFilmsDto.getFilmId()) {
-            directorDto.getFilmsIds();
+        if (orders.size() != 0) {
+            return;
         }
-        return (DirectorDto) directorRepository.findAll();
-    }*/
+        repository.deleteById(id);
+        Set<Film> singleDirectorFilms = filmRepository.findAllSingleDirectorFilmsByDirectorId(id);
 
+        for (Film film : singleDirectorFilms) {
+            filmRepository.deleteById(film.getId());
+        }
+    }
+
+    public Director addFilm(AddFilmDto addFilmDto) {
+        Director director = repository.findById(addFilmDto.getDirectorId()).orElseThrow();
+        Film film = filmRepository.findById(addFilmDto.getFilmId()).orElseThrow();
+        director.getFilms().add(film);
+        return repository.save(director);
+    }
+
+    public Page<Director> searchByDirectorFIO(Pageable pageable, String directorFIO) {
+        return repository.findAllByDirectorFIO(pageable, directorFIO);
+    }
+
+    public Page<Director> searchByDirectorFIOContaining(Pageable pageable, String directorFIO) {
+        return repository.findAllByDirectorFIOContaining(pageable, directorFIO);
+    }
+
+    public Page<Director> searchByDirectorFIOContainingForUsers(Pageable pageable, String directorFIO) {
+        return repository.findAllByDirectorFIOContainingAndIsDeletedFalse(pageable, directorFIO);
+    }
+
+    public Page<Director> listAllPaginated(Pageable pageable) {
+        return repository.findAll(pageable);
+    }
+
+    public Page<Director> listAllPaginatedForUsers(Pageable pageable) {
+        return repository.findAllByIsDeletedFalse(pageable);
+    }
+
+    public void block(Long id) {
+        Director director = getOne(id);
+        director.setDeleted(true);
+        director.setDeletedBy(
+                ((CustomUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal())
+                        .getUsername());
+        director.setDeletedWhen(LocalDateTime.now());
+        update(director);
+    }
+
+    public void unblock(Long id) {
+        Director director = getOne(id);
+        director.setDeleted(false);
+        director.setDeletedBy(null);
+        director.setDeletedWhen(null);
+        director.setUpdatedBy(
+                ((CustomUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal())
+                        .getUsername());
+        director.setUpdatedWhen(LocalDateTime.now());
+        update(director);
+    }
 }
