@@ -1,11 +1,11 @@
 package com.example.filmography.service;
 
-import com.example.filmography.model.Film;
+
+
 import com.example.filmography.model.GenericModel;
-import com.example.filmography.model.Genre;
 import com.example.filmography.repository.GenericRepository;
 import org.springframework.stereotype.Service;
-import org.webjars.NotFoundException;
+
 
 import java.util.List;
 
@@ -13,17 +13,18 @@ import java.util.List;
 public abstract class GenericService<T extends GenericModel> {
 
     private final GenericRepository<T> repository;
+
     @SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection")
     protected GenericService(GenericRepository<T> repository) {
         this.repository = repository;
     }
 
-    public List<T> getList() {
+    public List<T> listAll() {
         return repository.findAll();
     }
 
     public T getOne(Long id) {
-        return repository.findById(id).orElseThrow(() -> new NotFoundException("Row with such ID: " + id + " not found"));
+        return repository.findById(id).orElseThrow();
     }
 
     public T create(T object) {
